Removes Blake2s from algorithms and gadgets
