+++ algorithms/src/commitment/mod.rs
@@ -20,9 +20,6 @@ pub use bhp::*;
 pub mod bhp_compressed;
 pub use bhp_compressed::*;
 
-pub mod blake2s;
-pub use blake2s::*;
-
 pub mod pedersen;
 pub use pedersen::*;
