+++ gadgets/src/algorithms/commitment/tests.rs
@@ -21,16 +21,10 @@ use crate::{
     traits::{algorithms::CommitmentGadget, alloc::AllocGadget, FieldGadget},
 };
 use snarkvm_algorithms::{
-    commitment::{
-        BHPCommitment,
-        BHPCompressedCommitment,
-        Blake2sCommitment,
-        PedersenCommitment,
-        PedersenCompressedCommitment,
-    },
+    commitment::{BHPCommitment, BHPCompressedCommitment, PedersenCommitment, PedersenCompressedCommitment},
     CommitmentScheme,
 };
-use snarkvm_curves::edwards_bls12::{EdwardsProjective, Fq, Fr};
+use snarkvm_curves::edwards_bls12::{EdwardsProjective, Fq};
 use snarkvm_r1cs::{ConstraintSystem, TestConstraintSystem};
 use snarkvm_utilities::rand::UniformRand;
 
@@ -122,42 +116,4 @@ fn pedersen_compressed_commitment_gadget_test() {
             native_and_gadget_equivalence_test::<TestCommitment, TestCommitmentGadget>();
         assert_eq!(native_output, gadget_output.get_value().unwrap());
     }
-}
-
-#[test]
-fn blake2s_commitment_gadget_test() {
-    let mut cs = TestConstraintSystem::<Fr>::new();
-    let rng = &mut thread_rng();
-
-    let input: [u8; 32] = rng.gen();
-
-    let mut randomness = [0u8; 32];
-    rng.fill(&mut randomness);
-
-    let commitment = Blake2sCommitment::setup("blake2s_commitment_gadget_test");
-    let native_result = commitment.commit(&input, &randomness).unwrap();
-
-    let mut input_bytes = vec![];
-    for (byte_i, input_byte) in input.iter().enumerate() {
-        let cs = cs.ns(|| format!("input_byte_gadget_{}", byte_i));
-        input_bytes.push(UInt8::alloc(cs, || Ok(*input_byte)).unwrap());
-    }
-
-    let mut randomness_bytes = vec![];
-    for (byte_i, random_byte) in randomness.iter().enumerate() {
-        let cs = cs.ns(|| format!("randomness_byte_gadget_{}", byte_i));
-        randomness_bytes.push(UInt8::alloc(cs, || Ok(*random_byte)).unwrap());
-    }
-    let randomness_bytes = Blake2sRandomnessGadget(randomness_bytes);
-
-    let blake2s_gadget =
-        Blake2sCommitmentGadget::alloc_constant(&mut cs.ns(|| "gadget_parameters"), || Ok(&commitment)).unwrap();
-    let gadget_result = blake2s_gadget
-        .check_commitment_gadget(&mut cs.ns(|| "gadget_evaluation"), &input_bytes, &randomness_bytes)
-        .unwrap();
-
-    for (i, nr) in native_result.iter().enumerate() {
-        assert_eq!(*nr, gadget_result.0[i].value.unwrap());
-    }
-    assert!(cs.is_satisfied());
 }
