+++ gadgets/src/algorithms/prf/mod.rs
@@ -14,11 +14,5 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod blake2s;
-pub use blake2s::*;
-
 pub mod poseidon;
 pub use poseidon::*;
-
-#[cfg(test)]
-pub mod tests;
